Move boto3 clients to a different file to avoid circular imports
